Rename : 종류별 파일 및 디렉토리 분리
각 계층별로 브랜드/컬러/사이즈 분리
